+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModelTest.java
@@ -162,15 +162,15 @@ public class DefaultFilterComboBoxModelTest {
 	void filterWithSelection() {
 		testModel.filterSelectedItem().set(true);
 		testModel.setSelectedItem(BJORN);
-		testModel.visiblePredicate().set(item -> !item.equals(BJORN));
+		testModel.items().visiblePredicate().set(item -> !item.equals(BJORN));
 		assertEquals(NULL, testModel.getSelectedItem());
 		assertNull(testModel.selectionModel().selectedValue());
 
-		testModel.visiblePredicate().clear();
+		testModel.items().visiblePredicate().clear();
 		testModel.filterSelectedItem().set(false);
 		assertFalse(testModel.filterSelectedItem().get());
 		testModel.setSelectedItem(BJORN);
-		testModel.visiblePredicate().set(item -> !item.equals(BJORN));
+		testModel.items().visiblePredicate().set(item -> !item.equals(BJORN));
 		assertNotNull(testModel.getSelectedItem());
 		assertEquals(BJORN, testModel.selectionModel().selectedValue());
 	}
@@ -179,15 +179,15 @@ public class DefaultFilterComboBoxModelTest {
 	void visiblePredicate() {
 		testModel.addListDataListener(listDataListener);
 
-		testModel.visiblePredicate().set(item -> false);
+		testModel.items().visiblePredicate().set(item -> false);
 		assertEquals(1, testModel.getSize());
-		testModel.visiblePredicate().set(item -> true);
+		testModel.items().visiblePredicate().set(item -> true);
 		assertEquals(6, testModel.getSize());
-		testModel.visiblePredicate().set(item -> !item.equals(ANNA));
+		testModel.items().visiblePredicate().set(item -> !item.equals(ANNA));
 		assertEquals(5, testModel.getSize());
 		assertFalse(testModel.items().visible().contains(ANNA));
 		assertTrue(testModel.items().filtered().contains(ANNA));
-		testModel.visiblePredicate().set(item -> item.equals(ANNA));
+		testModel.items().visiblePredicate().set(item -> item.equals(ANNA));
 		assertEquals(2, testModel.getSize());
 		assertTrue(testModel.items().visible().contains(ANNA));
 
@@ -208,9 +208,9 @@ public class DefaultFilterComboBoxModelTest {
 	@Test
 	void remove() {
 		//remove filtered item
-		testModel.visiblePredicate().set(item -> !item.equals(BJORN));
+		testModel.items().visiblePredicate().set(item -> !item.equals(BJORN));
 		testModel.removeItem(BJORN);
-		testModel.visiblePredicate().clear();
+		testModel.items().visiblePredicate().clear();
 		assertFalse(testModel.items().visible().contains(BJORN));
 
 		//remove visible item
@@ -222,7 +222,7 @@ public class DefaultFilterComboBoxModelTest {
 	void add() {
 		testModel.clear();
 		//add filtered item
-		testModel.visiblePredicate().set(item -> !item.equals(BJORN));
+		testModel.items().visiblePredicate().set(item -> !item.equals(BJORN));
 		testModel.addItem(BJORN);
 		assertFalse(testModel.items().visible().contains(BJORN));
 
@@ -230,7 +230,7 @@ public class DefaultFilterComboBoxModelTest {
 		testModel.addItem(KALLI);
 		assertTrue(testModel.items().visible().contains(KALLI));
 
-		testModel.visiblePredicate().clear();
+		testModel.items().visiblePredicate().clear();
 		assertTrue(testModel.items().visible().contains(BJORN));
 	}